clean diffusion_module validation and move pocket_constraints to its own module
